rename lambda to set_container_overrides
