+++ apps/workflow-cf.yml.j2
@@ -103,7 +103,7 @@ Resources:
         {% endfor %}
         {% endfor %}
         UpdateDBLambdaArn: !GetAtt UpdateDB.Outputs.LambdaArn
-        OverrideMemoryLambdaArn: !GetAtt OverrideMemory.Outputs.LambdaArn
+        SetContainerOverridesLambdaArn: !GetAtt SetContainerOverrides.Outputs.LambdaArn
         GetFilesLambdaArn: !GetAtt GetFiles.Outputs.LambdaArn
         CheckProcessingTimeLambdaArn: !GetAtt CheckProcessingTime.Outputs.LambdaArn
         UploadLogLambdaArn: !GetAtt UploadLog.Outputs.LambdaArn
@@ -163,7 +163,7 @@ Resources:
             Action: lambda:InvokeFunction
             Resource:
              - !GetAtt UpdateDB.Outputs.LambdaArn
-             - !GetAtt OverrideMemory.Outputs.LambdaArn
+             - !GetAtt SetContainerOverrides.Outputs.LambdaArn
              - !GetAtt GetFiles.Outputs.LambdaArn
              - !GetAtt CheckProcessingTime.Outputs.LambdaArn
              - !GetAtt UploadLog.Outputs.LambdaArn
@@ -179,7 +179,7 @@ Resources:
         {% endif %}
       TemplateURL: update-db/update-db-cf.yml
 
-  OverrideMemory:
+  SetContainerOverrides:
     Type: AWS::CloudFormation::Stack
     Properties:
       {% if security_environment == 'EDC' %}
@@ -187,7 +187,7 @@ Resources:
         SecurityGroupId: !Ref SecurityGroupId
         SubnetIds: !Join [",", !Ref SubnetIds]
       {% endif %}
-      TemplateURL: override-memory/override-memory-cf.yml
+      TemplateURL: set-container-overrides/set-container-overrides-cf.yml
 
   GetFiles:
     Type: AWS::CloudFormation::Stack
